Update meter_page : Seperated meter_adview widget
